requirements.txt
pandas
numpy
tensorflow
python-jobspy

# docker pull tensorflow/tensorflow:latest
# docker run -it -p 8888:8888 tensorflow/tensorflow:latest-jupyter
